update shell and debug commands to follow new hierarchy pattern
