+++ webpage/src/CLI/lib/commands/debug/index.ts
@@ -1,0 +1,5 @@
+import { clearFS } from "./clearfs"
+
+export const debugCommands = {
+  "clearFs": clearFS,
+}
